refactor(wizard): split integration_validators/client_validators.py into per-vendor-family modules
client_validators.py had grown to 624 lines via the same per-vendor
accretion pattern as _integration_configurators.py (#4066/#4069): ~20
standalone validate_<vendor>_integration() functions with no shared
state beyond the IntegrationHealthResult type. Split into
observability.py, alerting.py, gitlab.py, posthog.py, sentry.py,
productivity.py, vercel.py, aws.py, dagster.py, and jenkins.py,
mirroring the vendor-family grouping already used in
surfaces/cli/wizard/configurators/.

Unlike _integration_configurators.py, this module had no unique
dispatcher logic to keep behind a facade, so client_validators.py is
removed outright and integration_health.py (the existing "stable
import surface") now imports each validator directly from its new
per-vendor-family module.

tests/cli/wizard/test_integration_health.py's monkeypatch string
targets (e.g. "...client_validators.DatadogClient") are updated to the
module that now owns each client import.

Fixes #4067
