+++ surfaces/cli/wizard/integration_health.py
@@ -2,28 +2,16 @@
 
 from __future__ import annotations
 
-from surfaces.cli.wizard.integration_validators.client_validators import (
+from surfaces.cli.wizard.integration_validators.alerting import (
     validate_alertmanager_integration,
-    validate_aws_integration,
     validate_betterstack_integration,
-    validate_coralogix_integration,
-    validate_dagster_integration,
-    validate_datadog_integration,
-    validate_gitlab_integration,
-    validate_google_docs_integration,
-    validate_grafana_integration,
-    validate_honeycomb_integration,
     validate_incident_io_integration,
-    validate_jenkins_integration,
-    validate_opensearch_integration,
     validate_opsgenie_integration,
     validate_pagerduty_integration,
-    validate_posthog_integration,
-    validate_sentry_integration,
-    validate_splunk_integration,
-    validate_tempo_integration,
-    validate_vercel_integration,
 )
+from surfaces.cli.wizard.integration_validators.aws import validate_aws_integration
+from surfaces.cli.wizard.integration_validators.dagster import validate_dagster_integration
+from surfaces.cli.wizard.integration_validators.gitlab import validate_gitlab_integration
 from surfaces.cli.wizard.integration_validators.http_probe_validators import (
     validate_discord_bot,
     validate_jira_integration,
@@ -31,13 +19,29 @@ from surfaces.cli.wizard.integration_validators.http_probe_validators import (
     validate_slack_webhook,
     validate_telegram_bot,
 )
+from surfaces.cli.wizard.integration_validators.jenkins import validate_jenkins_integration
 from surfaces.cli.wizard.integration_validators.mcp_validators import (
     validate_github_mcp_integration,
     validate_openclaw_integration,
     validate_posthog_mcp_integration,
     validate_sentry_mcp_integration,
 )
+from surfaces.cli.wizard.integration_validators.observability import (
+    validate_coralogix_integration,
+    validate_datadog_integration,
+    validate_grafana_integration,
+    validate_honeycomb_integration,
+    validate_opensearch_integration,
+    validate_splunk_integration,
+    validate_tempo_integration,
+)
+from surfaces.cli.wizard.integration_validators.posthog import validate_posthog_integration
+from surfaces.cli.wizard.integration_validators.productivity import (
+    validate_google_docs_integration,
+)
+from surfaces.cli.wizard.integration_validators.sentry import validate_sentry_integration
 from surfaces.cli.wizard.integration_validators.shared import IntegrationHealthResult
+from surfaces.cli.wizard.integration_validators.vercel import validate_vercel_integration
 
 __all__ = [
     "IntegrationHealthResult",
