+++ tests/cli/wizard/test_integration_health.py
@@ -86,7 +86,7 @@ class _FakeDatadogClient:
 
 def test_validate_grafana_integration_succeeds_when_datasources_are_discovered(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.get_grafana_client_from_credentials",
+        "surfaces.cli.wizard.integration_validators.observability.get_grafana_client_from_credentials",
         lambda **_kwargs: _FakeGrafanaClient({"loki_uid": "loki-1", "tempo_uid": "tempo-1"}),
     )
 
@@ -98,7 +98,7 @@ def test_validate_grafana_integration_succeeds_when_datasources_are_discovered(m
 
 def test_validate_grafana_integration_fails_when_no_datasources_are_found(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.get_grafana_client_from_credentials",
+        "surfaces.cli.wizard.integration_validators.observability.get_grafana_client_from_credentials",
         lambda **_kwargs: _FakeGrafanaClient({}),
     )
 
@@ -110,7 +110,7 @@ def test_validate_grafana_integration_fails_when_no_datasources_are_found(monkey
 
 def test_validate_datadog_integration_succeeds(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.DatadogClient",
+        "surfaces.cli.wizard.integration_validators.observability.DatadogClient",
         lambda _config: _FakeDatadogClient({"success": True, "total": 7}),
     )
 
@@ -122,7 +122,7 @@ def test_validate_datadog_integration_succeeds(monkeypatch) -> None:
 
 def test_validate_datadog_integration_fails(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.DatadogClient",
+        "surfaces.cli.wizard.integration_validators.observability.DatadogClient",
         lambda _config: _FakeDatadogClient({"success": False, "error": "HTTP 403"}),
     )
 
@@ -134,11 +134,11 @@ def test_validate_datadog_integration_fails(monkeypatch) -> None:
 
 def test_validate_honeycomb_integration_succeeds(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.HoneycombClient.validate_access",
+        "surfaces.cli.wizard.integration_validators.observability.HoneycombClient.validate_access",
         lambda _self: {"success": True, "environment": {"slug": "prod"}},
     )
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.HoneycombClient.run_query",
+        "surfaces.cli.wizard.integration_validators.observability.HoneycombClient.run_query",
         lambda _self, *_args, **_kwargs: {"success": True, "results": [{}]},
     )
 
@@ -167,7 +167,7 @@ def test_validate_incident_io_integration_succeeds(monkeypatch) -> None:
             return {"success": True}
 
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.IncidentIoClient",
+        "surfaces.cli.wizard.integration_validators.alerting.IncidentIoClient",
         _FakeIncidentIoClient,
     )
 
@@ -179,7 +179,7 @@ def test_validate_incident_io_integration_succeeds(monkeypatch) -> None:
 
 def test_validate_coralogix_integration_fails(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.CoralogixClient.validate_access",
+        "surfaces.cli.wizard.integration_validators.observability.CoralogixClient.validate_access",
         lambda _self: {"success": False, "error": "HTTP 401"},
     )
 
@@ -354,7 +354,7 @@ def test_validate_github_mcp_integration_uses_shared_validator(monkeypatch) -> N
 
 def test_validate_sentry_integration_uses_shared_validator(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.validate_sentry_config",
+        "surfaces.cli.wizard.integration_validators.sentry.validate_sentry_config",
         lambda _config: types.SimpleNamespace(ok=True, detail="Sentry ok"),
     )
 
@@ -371,7 +371,7 @@ def test_validate_sentry_integration_uses_shared_validator(monkeypatch) -> None:
 
 def test_validate_dagster_integration_uses_shared_validator(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.validate_dagster_config",
+        "surfaces.cli.wizard.integration_validators.dagster.validate_dagster_config",
         lambda _config: types.SimpleNamespace(
             ok=True, detail="Connected to Dagster version mock-1.0."
         ),
@@ -388,7 +388,7 @@ def test_validate_dagster_integration_uses_shared_validator(monkeypatch) -> None
 
 def test_validate_dagster_integration_surfaces_failure(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.validate_dagster_config",
+        "surfaces.cli.wizard.integration_validators.dagster.validate_dagster_config",
         lambda _config: types.SimpleNamespace(
             ok=False, detail="Dagster GraphQL probe failed: HTTP 401"
         ),
@@ -419,7 +419,7 @@ class _FakeVercelClient:
 
 def test_validate_vercel_integration_succeeds(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.VercelClient",
+        "surfaces.cli.wizard.integration_validators.vercel.VercelClient",
         lambda _config: _FakeVercelClient(
             {"success": True, "projects": [{"id": "p1"}], "total": 1}
         ),
@@ -448,7 +448,7 @@ def test_validate_vercel_integration_succeeds_with_team_id(monkeypatch) -> None:
             return {"success": True, "projects": [], "total": 0}
 
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.VercelClient",
+        "surfaces.cli.wizard.integration_validators.vercel.VercelClient",
         _CapturingClient,
     )
 
@@ -460,7 +460,7 @@ def test_validate_vercel_integration_succeeds_with_team_id(monkeypatch) -> None:
 
 def test_validate_vercel_integration_fails_on_api_error(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.VercelClient",
+        "surfaces.cli.wizard.integration_validators.vercel.VercelClient",
         lambda _config: _FakeVercelClient({"success": False, "error": "HTTP 401: unauthorized"}),
     )
 
@@ -482,7 +482,7 @@ def test_validate_vercel_integration_surfaces_exception(monkeypatch) -> None:
         raise RuntimeError("network unreachable")
 
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.VercelClient",
+        "surfaces.cli.wizard.integration_validators.vercel.VercelClient",
         _raise,
     )
 
@@ -614,7 +614,7 @@ def test_validate_telegram_bot_network_error(monkeypatch: pytest.MonkeyPatch) ->
 
 def test_validate_betterstack_integration_succeeds(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.validate_betterstack_config",
+        "surfaces.cli.wizard.integration_validators.alerting.validate_betterstack_config",
         lambda _config: BetterStackValidationResult(ok=True, detail="Connected."),
     )
     result = validate_betterstack_integration(
@@ -629,7 +629,7 @@ def test_validate_betterstack_integration_succeeds(monkeypatch) -> None:
 
 def test_validate_betterstack_integration_forwards_failure_detail(monkeypatch) -> None:
     monkeypatch.setattr(
-        "surfaces.cli.wizard.integration_validators.client_validators.validate_betterstack_config",
+        "surfaces.cli.wizard.integration_validators.alerting.validate_betterstack_config",
         lambda _config: BetterStackValidationResult(
             ok=False, detail="Better Stack authentication failed."
         ),
